* Modularize main.rs with clean separation of concerns   - Remove inline module definitions for better code organization   - Use placeholder modules for development and testing   - Maintain zero-copy performance optimizations throughout
