+++ drv/openprot-ecdsa-server/src/placeholder.rs
@@ -1,0 +1,105 @@
+//! Placeholder implementations for ECDSA operations
+//!
+//! This module provides concrete placeholder implementations of the OpenPRoT
+//! ECDSA traits for testing and development purposes. These implementations
+//! provide the correct trait interfaces but do not perform actual cryptographic
+//! operations.
+
+use crate::p384_key;
+use openprot_hal_blocking::ecdsa::{EcdsaVerify, ErrorType, EcdsaSign, PrivateKey, P384, ErrorKind};
+use openprot_hal_blocking::digest::DigestAlgorithm;
+use zeroize::{Zeroize, ZeroizeOnDrop};
+
+/// Placeholder ECDSA Verifier Implementation
+/// 
+/// This is a concrete type that implements the EcdsaVerify trait for P384.
+/// It provides a minimal working implementation for testing and development.
+pub struct PlaceholderVerifier;
+
+impl ErrorType for PlaceholderVerifier {
+    type Error = core::convert::Infallible;
+}
+
+impl EcdsaVerify<P384> for PlaceholderVerifier {
+    type PublicKey = p384_key::P384PublicKey;
+    type Signature = p384_key::P384Signature;
+    
+    fn verify(
+        &mut self,
+        _public_key: &Self::PublicKey,
+        _digest: <<P384 as openprot_hal_blocking::ecdsa::Curve>::DigestType as DigestAlgorithm>::Digest,
+        _signature: &Self::Signature,
+    ) -> Result<(), Self::Error> {
+        // Placeholder implementation - always "succeeds" 
+        // In a real implementation, this would:
+        // 1. Extract the r,s components from the signature
+        // 2. Extract the x,y coordinates from the public key  
+        // 3. Perform the ECDSA verification algorithm on P384 curve
+        // 4. Return Ok(()) if verification passes, Err(error) if it fails
+        Ok(())
+    }
+}
+
+/// P384 Private Key Implementation
+/// 
+/// This concrete type implements both Zeroize and PrivateKey<P384> traits
+/// as required by the OpenPRoT framework.
+#[derive(Zeroize, ZeroizeOnDrop)]
+#[repr(C)]
+pub struct P384PrivateKey {
+    /// Private scalar value (48 bytes for P384)
+    scalar: [u8; 48],
+}
+
+impl P384PrivateKey {
+    /// Create a new private key from scalar bytes
+    pub fn new(scalar: [u8; 48]) -> Self {
+        Self { scalar }
+    }
+}
+
+impl PrivateKey<P384> for P384PrivateKey {
+    fn validate(&self) -> Result<(), ErrorKind> {
+        // TODO: Add proper private key validation
+        // For now, accept any 48-byte value as valid
+        // In a real implementation, validate that:
+        // 1 ≤ scalar < curve_order
+        Ok(())
+    }
+}
+
+/// Placeholder ECDSA Signer Implementation
+/// 
+/// This is a concrete type that implements the EcdsaSign trait for P384.
+/// It provides a minimal working implementation for testing and development.
+pub struct PlaceholderSigner;
+
+impl ErrorType for PlaceholderSigner {
+    type Error = core::convert::Infallible;
+}
+
+impl EcdsaSign<P384> for PlaceholderSigner {
+    type PrivateKey = P384PrivateKey; // Use our concrete private key type
+    type Signature = p384_key::P384Signature;
+    
+    fn sign<R>(
+        &mut self,
+        _private_key: &Self::PrivateKey,
+        _digest: <<P384 as openprot_hal_blocking::ecdsa::Curve>::DigestType as DigestAlgorithm>::Digest,
+        _rng: &mut R,
+    ) -> Result<Self::Signature, Self::Error> {
+        // Placeholder implementation - returns a dummy signature
+        // In a real implementation, this would:
+        // 1. Validate the private key
+        // 2. Use the digest and private key to generate ECDSA signature components (r, s)
+        // 3. Use random number generator for signature generation
+        // 4. Return the computed signature
+        
+        // Create a deterministic placeholder signature for testing
+        let placeholder_r = [0u8; 48];
+        let mut placeholder_s = [1u8; 48];
+        placeholder_s[47] = 42; // Make it slightly non-trivial
+        
+        Ok(p384_key::P384Signature::new(placeholder_r, placeholder_s))
+    }
+}
